разделила регистрацию на две странице
