+++ src/pages/register-forn/styles.module.css
@@ -1,0 +1,27 @@
+.header {
+  height: 90px;
+}
+.titlePrimary {
+  margin: 0;
+  padding: 40px 18px 12px;
+  font-family: 'Golos', sans-serif;
+  font-size: 24px;
+  line-height: 29px;
+  color: var(--colors-interface-primary);
+}
+.form {
+  min-width: 364px;
+  display: grid;
+  gap: 24px;
+  padding-left: 18px;
+  justify-items: flex-start;
+}
+.field {
+  width: 364px;
+}
+.text {
+  font-family: 'Golos', sans-serif;
+  font-size: 12px;
+  line-height: 14px;
+  padding-bottom: 6px;
+}
